Use ObjectSummary.size instead of .content_length to get entry size
The S3 boto docs are missing an explanation for what .size actually is, but assuming it works the same as in the Ruby SDK these two attributes should both represent the object size.

This patch fixes the problem where trying to retrieve the size of an object in the local cache throws an error: 's3.ObjectSummary' object has no attribute 'content_length'.
It occurs because the items in the entries dictionary are of type ObjectSummary instead of Object (the latter has a .content_length attribute, the former only has .size).